Created extension for table view header layout
