+++ WordPress/Classes/ViewRelated/Site Creation/SiteInfo/SiteInformationWizardContent.swift
@@ -67,20 +67,9 @@ final class SiteInformationWizardContent: UIViewController {
         stopListeningToKeyboardNotifications()
     }
 
-    // via https://collindonnell.com/2015/09/29/dynamically-sized-table-view-header-or-footer-using-auto-layout/
     override func viewDidLayoutSubviews() {
         super.viewDidLayoutSubviews()
-
-        if let headerView = table.tableHeaderView {
-            let height = headerView.systemLayoutSizeFitting(UIView.layoutFittingCompressedSize).height
-            var headerFrame = headerView.frame
-
-            if height != headerFrame.size.height {
-                headerFrame.size.height = height
-                headerView.frame = headerFrame
-                table.tableHeaderView = headerView
-            }
-        }
+        table.layoutHeaderView()
     }
 
     private func applyTitle() {
